+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: msokolov <msokolov@student.codam.nl>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/16 17:42:44 by msokolov          #+#    #+#             */
-/*   Updated: 2025/12/04 12:48:35 by msokolov         ###   ########.fr       */
+/*   Updated: 2025/12/04 22:47:49 by msokolov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -80,7 +80,7 @@ t_list	*ft_lstmap(t_list *lst, void *(*f)(void *), void (*del)(void *));
 t_list	*ft_lstlast(t_list *lst);
 
 size_t	ft_strlen1(char *s);
-size_t	ft_strlen(const char *c);
+int		ft_strlen(const char *c);
 size_t	ft_strlcpy(char *dest, const char *src, size_t size);
 size_t	ft_strlcat(char *dest, const char *src, size_t size);
 #endif
